+++ src/Client/State/Focus.hs
@@ -50,6 +50,7 @@ data Subfocus
   | FocusWindows WindowsFilter -- ^ Show client windows
   | FocusPalette     -- ^ Show current palette
   | FocusMentions    -- ^ Show all mentions
+  | FocusDigraphs    -- ^ Show all digraphs
   | FocusHelp (Maybe Text) -- ^ Show help window with optional command
   deriving (Eq,Show)